+++ app/src/main/java/yandex/school/project/presentation/common/Result.kt
@@ -1,7 +1,7 @@
 package yandex.school.project.presentation.common
 
 sealed class Result<out T> {
-    object Loading : Result<Nothing>()
+    data object Loading : Result<Nothing>()
     data class Success<T>(val data: T) : Result<T>()
     data class Error(val message: String) : Result<Nothing>()
 }
