+++ app/src/main/java/yandex/school/project/presentation/screens/MainScreen.kt
@@ -5,15 +5,12 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
-import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -22,19 +19,11 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.navigation.compose.rememberNavController
+import yandex.school.project.presentation.components.TopBar
+import yandex.school.project.presentation.components.TopBarState
 import yandex.school.project.presentation.navigation.BottomBar
 import yandex.school.project.presentation.navigation.BottomNavigation
 import yandex.school.project.presentation.theme.ProjectTheme
-
-data class TopBarState(
-    val navigationIcon: ImageVector? = null,
-    val navigationIconAction: () -> Unit = {},
-    val title: String,
-    val actionIcon: ImageVector? = null,
-    val actionIconAction: () -> Unit = {},
-    val isFAB: Boolean = false
-)
-
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -53,41 +42,7 @@ fun MainScreen(accountId: Int) {
 
     Scaffold(
         topBar = {
-            CenterAlignedTopAppBar(
-                navigationIcon = {
-                    if (currentTopBarState.value.navigationIcon != null) {
-                        IconButton(
-                            onClick = currentTopBarState.value.navigationIconAction
-                        ) {
-                            Icon(
-                                imageVector = currentTopBarState.value.navigationIcon!!,
-                                null
-                            )
-                        }
-                    }
-                },
-                title = {
-                    Text(
-                        text = currentTopBarState.value.title,
-                        style = MaterialTheme.typography.titleLarge
-                    )
-                },
-                actions = {
-                    if (currentTopBarState.value.actionIcon != null) {
-                        IconButton(
-                            onClick = currentTopBarState.value.actionIconAction
-                        ) {
-                            Icon(
-                                imageVector = currentTopBarState.value.actionIcon!!,
-                                null
-                            )
-                        }
-                    }
-                },
-                colors = TopAppBarDefaults.topAppBarColors(
-                    containerColor = MaterialTheme.colorScheme.primary
-                )
-            )
+            TopBar(currentTopBarState.value)
         },
         bottomBar = { BottomBar(bottomNavController) },
         floatingActionButton = {
