+++ app/src/main/java/yandex/school/project/presentation/screens/category/CategoryScreen.kt
@@ -1,20 +1,13 @@
 package yandex.school.project.presentation.screens.category
 
-import android.widget.Toast
 import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Search
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.Divider
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
@@ -23,28 +16,22 @@ import androidx.compose.material3.OutlinedTextFieldDefaults
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.RectangleShape
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.viewmodel.compose.viewModel
-import yandex.school.project.presentation.common.Result
-import yandex.school.project.presentation.components.ErrorItem
 import yandex.school.project.presentation.components.ListItem
-import yandex.school.project.presentation.screens.category.CategoryViewModel
+import yandex.school.project.presentation.components.ResultScreen
 import yandex.school.project.presentation.theme.ProjectTheme
 
-@OptIn(ExperimentalMaterial3Api::class, ExperimentalFoundationApi::class)
+@OptIn(ExperimentalFoundationApi::class)
 @Composable
 fun CategoryScreen(
     viewModel: CategoryViewModel = hiltViewModel()
@@ -52,57 +39,46 @@ fun CategoryScreen(
     var search by remember { mutableStateOf("") }
     val uiState by viewModel.uiState.collectAsState()
 
-    when (val state = uiState) {
-        is Result.Loading -> {
-            Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
-                CircularProgressIndicator()
-            }
-        }
-        is Result.Error -> {
-            ErrorItem(
-                errorMessage = state.message
-            ) {
-                viewModel.loadCategoriesWithRetry()
-            }
-        }
-        is Result.Success -> {
-            val categories = state.data.filter { it.name.contains(search, ignoreCase = true) }
-            HorizontalDivider()
-            LazyColumn(modifier = Modifier.fillMaxSize()) {
-                stickyHeader {
-                    OutlinedTextField(
-                        value = search,
-                        onValueChange = { search = it },
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .height(56.dp),
-                        placeholder = { Text("Найти статью") },
-                        trailingIcon = {
-                            Icon(imageVector = Icons.Default.Search, contentDescription = null)
-                        },
-                        singleLine = true,
-                        shape = RectangleShape,
-                        colors = OutlinedTextFieldDefaults.colors(
-                            unfocusedTextColor = MaterialTheme.colorScheme.onSurfaceVariant,
-                            focusedContainerColor = MaterialTheme.colorScheme.surfaceContainerHigh,
-                            unfocusedContainerColor = MaterialTheme.colorScheme.surfaceContainerHigh,
-                            disabledContainerColor = MaterialTheme.colorScheme.surfaceContainerHigh,
-                            errorContainerColor = Color.Transparent,
-                            focusedBorderColor = Color.Transparent,
-                            unfocusedBorderColor = Color.Transparent,
-                            unfocusedTrailingIconColor = MaterialTheme.colorScheme.onSurfaceVariant,
-                        )
+    ResultScreen(
+        result = uiState,
+        onRetry = { viewModel.loadCategoriesWithRetry() }
+    ) { categories ->
+        val filtered = categories.filter { it.name.contains(search, ignoreCase = true) }
+        HorizontalDivider()
+        LazyColumn(modifier = Modifier.fillMaxSize()) {
+            stickyHeader {
+                OutlinedTextField(
+                    value = search,
+                    onValueChange = { search = it },
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .height(56.dp),
+                    placeholder = { Text("Найти статью") },
+                    trailingIcon = {
+                        Icon(imageVector = Icons.Default.Search, contentDescription = null)
+                    },
+                    singleLine = true,
+                    shape = RectangleShape,
+                    colors = OutlinedTextFieldDefaults.colors(
+                        unfocusedTextColor = MaterialTheme.colorScheme.onSurfaceVariant,
+                        focusedContainerColor = MaterialTheme.colorScheme.surfaceContainerHigh,
+                        unfocusedContainerColor = MaterialTheme.colorScheme.surfaceContainerHigh,
+                        disabledContainerColor = MaterialTheme.colorScheme.surfaceContainerHigh,
+                        errorContainerColor = Color.Transparent,
+                        focusedBorderColor = Color.Transparent,
+                        unfocusedBorderColor = Color.Transparent,
+                        unfocusedTrailingIconColor = MaterialTheme.colorScheme.onSurfaceVariant,
                     )
-                    HorizontalDivider()
-                }
-                items(categories) { category ->
-                    ListItem(
-                        modifier = Modifier.height(70.dp),
-                        leadingIcon = category.icon ?: "📁",
-                        contentTitle = category.name
-                    )
-                    HorizontalDivider()
-                }
+                )
+                HorizontalDivider()
+            }
+            items(filtered) { category ->
+                ListItem(
+                    modifier = Modifier.height(70.dp),
+                    leadingIcon = category.icon ?: "📁",
+                    contentTitle = category.name
+                )
+                HorizontalDivider()
             }
         }
     }
